Basic structure of CSS based navigation
This is the basic structure of the CSS navigation. The slide out menus
are not populated with their content in this version but it operates
properly. I was hoping to not have to do much more than just plug in
the inner content of the slide out menus but it is proving to be not
that simple. I am working through that and hope to have it all
functioning properly soon.