Updated full demo site to CSS based slide outs
This index.html file is the full demo site using the CSS based
navigation and slide-out menus and only contains the necessary jquery
for what is on the site (old jquery used for the slide-out menus has
been removed). The styles.css file has also been updated to contain the
necessary css for the new slide-out menu.