Delete Tac Button, Board Name on Timeline
Added delete tac button. This can be found on the timeline. Seemed like
the quickest and easiest place to put it for now.
Added the boards name to tags found on the timeline. These will appear
when the user select "All Boards" or "My Tacs" from the timeline. They
aren't necessary when selecting "This Board"